pk_internal_tools/pk_web_server/pyproject.toml
[project]
name = "pk_web_server"
version = "0.1.0"
description = "Independent FastAPI web server for pk_system"
dependencies = [
    "fastapi>=0.104.1",
    "uvicorn>=0.24.0",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "pytest>=8.3.4", # 테스트 프레임워크
    "pytest-asyncio>=0.23.6", # 비동기 테스트 지원
    "httpx>=0.27.0", # FastAPI 테스트 클라이언트
]
requires-python = ">=3.10"
license = { text = "MIT" }

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.uv]
# requires-python은 [project] 섹션에 정의되어야 합니다.
# 이 섹션은 uv 관련 추가 설정을 정의할 수 있습니다.
